Adjustments to work with tags and equity.
This ensures that the cli works with
https://github.com/PyconUK/ConferenceScheduler/pull/103

I have tested various objective functions and also tested the validator.

Note I needed to adjust the tags: as was it was not working as it was
loading all tags as `''` which essentially meant it was not possible for
the scheduler to find a solution (as all talks had the same tag so where
all unavailable during each other).